Move `MarketStrategy` and subclasses to a separate module
